requirements.txt
aiogram
asyncio
python-dotenv
